requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.1.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-taggit==6.0.0
environs==11.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.9
import-export==0.3.1
marshmallow==3.22.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.1
stripe==10.11.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
